automations.yaml: try upstairs fan 3x
The wifi connection for the upstairs fan is pretty flakey, so it often goes
offline for a short while.  If it just happens to be offline as the
automation triggers then the state change is lost and the fan is kept
running/stopped.

As a workaround, trigger the automation 3x, 5 minutes apart to be (more)
sure the state change takes effect.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>